rename colors namespace to Colors
