+++ src/Colors/Lch.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace Atomicptr\Color\colors;
+namespace Atomicptr\Color\Colors;
 
 use       Atomicptr\Color\Color;
 use       Atomicptr\Color\ColorFactory;
 use       Atomicptr\Color\ColorInterface;
 use       Atomicptr\Color\utils;
 
-class      Hwb
+class      Lch
 extends    Color
 implements ColorInterface {
 
     /* #region Constructor */
 
     public function __construct(
+        public readonly float $lightness = 0,
+        public readonly float $chroma    = 0,
         public readonly float $hue       = 0,
-        public readonly float $whiteness = 0,
-        public readonly float $blackness = 0,
         public readonly float $opacity   = 100,
     ) {
 
@@ -30,7 +30,10 @@ implements ColorInterface {
 
     ) :array {
         return [
-            'hwb',
+            'lch',
+            'cielch',
+            'cie-lch',
+            'cie_lch',
         ];
     }
 
@@ -39,20 +42,20 @@ implements ColorInterface {
     /* #region Public Methods */
 
     public function change(
+        \Stringable|string|int|float|null $lightness = null,
+        \Stringable|string|int|float|null $chroma    = null,
         \Stringable|string|int|float|null $hue       = null,
-        \Stringable|string|int|float|null $whiteness = null,
-        \Stringable|string|int|float|null $blackness = null,
         \Stringable|string|int|float|null $opacity   = null,
-        Hwb|null                          $fallback  = null,
+        Lch|null                          $fallback  = null,
         bool|null                         $throw     = null,
-    ) :Hwb {
+    ) :Lch {
         $changeThrow = $throw ?? true;
 
-        return ColorFactory::newHwb(
+        return ColorFactory::newLch(
             value    : [
+                utils\changeCoordinate($this->lightness, $lightness, false, $changeThrow),
+                utils\changeCoordinate($this->chroma,    $chroma,    false, $changeThrow),
                 utils\changeCoordinate($this->hue,       $hue,       false, $changeThrow),
-                utils\changeCoordinate($this->whiteness, $whiteness, false, $changeThrow),
-                utils\changeCoordinate($this->blackness, $blackness, false, $changeThrow),
                 utils\changeCoordinate($this->opacity,   $opacity,   false, $changeThrow),
             ],
             from     : $this::space(),
@@ -66,10 +69,10 @@ implements ColorInterface {
         bool|null $alpha     = null,
         int|null  $precision = null,
     ) :string {
-        return utils\hwb\stringify(
+        return utils\lch\stringify(
+            lightness : $this->lightness,
+            chroma    : $this->chroma,
             hue       : $this->hue,
-            whiteness : $this->whiteness,
-            blackness : $this->blackness,
             opacity   : $this->opacity,
             legacy    : $legacy,
             alpha     : $alpha,
@@ -78,4 +81,5 @@ implements ColorInterface {
     }
 
     /* #endregion */
+
 }
